requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
APScheduler==3.11.0
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
discord.py==2.5.2
dotenv==0.9.9
frozenlist==1.6.2
idna==3.10
instaloader==4.14.1
multidict==6.4.4
propcache==0.3.1
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
tzlocal==5.3.1
urllib3==2.4.0
yarl==1.20.0
